+++ server/form-pages/apply/offence-information/previous-unspent-convictions/index.ts
@@ -6,8 +6,8 @@ import OffenceHistoryData from './custom-forms/offenceHistoryData'
 import OffenceHistory from './offenceHistory'
 
 @Task({
-  name: 'Add offending history',
-  slug: 'offending-history',
+  name: 'Add previous unspent convictions',
+  slug: 'previous-unspent-convictions',
   pages: [AnyPreviousConvictions, OffenceHistoryData, OffenceHistory],
 })
-export default class OffendingHistory {}
+export default class PreviousUnspentConvictions {}
