+++ server/form-pages/utils/questions.ts
@@ -927,13 +927,13 @@ export const getQuestions = (name: string) => {
         },
       },
     },
-    'offending-history': {
+    'previous-unspent-convictions': {
       'any-previous-convictions': {
         hasAnyPreviousConvictions: {
           question: `Does ${name} have any previous unspent convictions?`,
           answers: {
-            yesRelevantRisk: 'Yes, and there is relevant risk',
-            yesNoRelevantRisk: 'Yes, but there is no relevant risk',
+            yesRelevantRisk: 'Yes and it is from one of the relevant categories',
+            yesNoRelevantRisk: 'Yes and it is not from one of the relevant categories',
             no: 'No, they do not have any previous unspent convictions',
           },
         },
